requirements.txt
pandas==2.2.3
numpy==2.2.4
matplotlib==3.10.1
seaborn==0.13.2
scipy==1.15.2
streamlit==1.44.1
requests==2.32.3
python-dotenv==1.1.0
sqlite3-binary==0.0.3
